requirements.txt
argparse
pandas
pathlib
requests
urllib3
pyyaml